+++ tests/Unit/SimpleSchedulerTest.php
@@ -6,6 +6,7 @@ use Closure;
 use Cron\CronExpression;
 use DateTimeImmutable;
 use DateTimeZone;
+use ErrorException;
 use Exception;
 use Generator;
 use Orisai\Clock\FrozenClock;
@@ -31,6 +32,9 @@ use Tests\Orisai\Scheduler\Doubles\JobInnerFailure;
 use Tests\Orisai\Scheduler\Doubles\TestLockFactory;
 use Throwable;
 use function rtrim;
+use function set_error_handler;
+use const E_ALL;
+use const E_USER_NOTICE;
 
 final class SimpleSchedulerTest extends TestCase
 {
@@ -1063,6 +1067,84 @@ Problem: Job subprocess produced stderr output.
 stderr: job error
 MSG,
 				rtrim($suppressed->getMessage()),
+			);
+		}
+	}
+
+	/**
+	 * @runInSeparateProcess
+	 */
+	public function testProcessJobStdout(): void
+	{
+		$errors = [];
+
+		set_error_handler(
+			static function (int $errno, string $errstr) use (&$errors): bool {
+				$errors[] = [$errno, $errstr];
+
+				return true;
+			},
+			E_ALL,
+		);
+
+		$scheduler = SchedulerProcessSetup::createWithStdoutJob();
+		$summary = $scheduler->run();
+
+		self::assertCount(1, $summary->getJobSummaries());
+		self::assertCount(1, $errors);
+		foreach ($errors as [$code, $message]) {
+			self::assertSame(E_USER_NOTICE, $code);
+			self::assertStringMatchesFormat(
+				<<<'MSG'
+Context: Running job via command %a
+Problem: Job subprocess produced unsupported stdout output.
+stdout:  echo%c
+MSG,
+				$message,
+			);
+		}
+	}
+
+	/**
+	 * @runInSeparateProcess
+	 */
+	public function testProcessJobStdoutWithStrictErrorHandler(): void
+	{
+		set_error_handler(
+			static function (int $errno, string $errstr): void {
+				throw new ErrorException($errstr, $errno);
+			},
+			E_ALL,
+		);
+
+		$scheduler = SchedulerProcessSetup::createWithStdoutJob();
+
+		$e = null;
+		try {
+			$scheduler->run();
+		} catch (RunFailure $e) {
+			// Handled bellow
+		}
+
+		self::assertNotNull($e);
+		self::assertStringStartsWith(
+			<<<'MSG'
+Run failed
+Suppressed errors:
+MSG,
+			$e->getMessage(),
+		);
+
+		self::assertNotSame([], $e->getSuppressed());
+		foreach ($e->getSuppressed() as $suppressed) {
+			self::assertInstanceOf(ErrorException::class, $suppressed);
+			self::assertStringMatchesFormat(
+				<<<'MSG'
+Context: Running job via command %a
+Problem: Job subprocess produced unsupported stdout output.
+stdout:  echo%c
+MSG,
+				$suppressed->getMessage(),
 			);
 		}
 	}